+++ app/helpers/backoffice/services_helper.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Backoffice::ServicesHelper
+  def service_status(service)
+    status_badge_class = service.published? ? "badge-success" : "badge-warning"
+    content_tag(:span, service.status, class: "badge #{status_badge_class}")
+  end
+end